+++ src/aioway/io/codecs/images.py
