+++ tests/io/codecs/test_audios.py
@@ -5,7 +5,7 @@ import pathlib
 import pytest
 import torch
 
-from aioway.codecs import (
+from aioway.io import (
     AudioLoader,
     AvAudioLoader,
     TorchCodecAudioLoader,
