+++ tests/io/codecs/test_videos.py
@@ -5,7 +5,7 @@ import pathlib
 import pytest
 import torch
 
-from aioway.codecs import AvVideoLoader, TorchCodecVideoLoader, VideoLoader
+from aioway.io import AvVideoLoader, TorchCodecVideoLoader, VideoLoader
 
 
 def _loader():
